# Conflicts:
#	core/src/main/java/farm/com/GameState.java

+++ core/src/main/java/farm/com/GameState.java
@@ -1,7 +1,7 @@
 package farm.com;
 
 public class GameState {
-    public static float money = 100000;
+    public static int money = 100000;
     public static int milkTotal = 0;
     public static int egg = 0;
     public static int totalPork = 0;
